Migrate getDistance
