+++ shopsafe-backend/src/main/java/com/google/sps/data/StoreStats.java
@@ -27,6 +27,7 @@ public final class StoreStats extends Store {
     private final double hygiene;
     private final double masks;
     private final long checkInCount;
+    private final double distance;
 
     // Static final weights for calculating score.
     static private final double COUNTY_WEIGHT = 0.5;
@@ -40,19 +41,21 @@ public final class StoreStats extends Store {
         this.hygiene = stats.getHygiene();
         this.masks = stats.getMasks();
         this.checkInCount = stats.getCheckInCount();
+        this.distance = 0; // default value
     }
 
     /**
      * Overloaded constructor to include userLocation to calculate distance between user and store
      */
     public StoreStats(Store store, double countyScore, CheckInStats stats, LatLng userLocation) {
-        super(store.id ,store.name, store.address, store.open, new LatLng(store.latitude, store.longitude), userLocation, store.rating);
+        super(store.id ,store.name, store.address, store.open, new LatLng(store.latitude, store.longitude), store.rating);
         this.score = countyScore * COUNTY_WEIGHT + stats.getCheckInScore() * CHECK_IN_WEIGHT;
         this.busy = stats.getBusy();
         this.line = stats.getLine();
         this.hygiene = stats.getHygiene();
         this.masks = stats.getMasks();
         this.checkInCount = stats.getCheckInCount();
+        this.distance = getDistance(userLocation);
     }
 
     public String getName() {
@@ -105,5 +108,38 @@ public final class StoreStats extends Store {
 
     public double getDistance() {
         return distance;
+    }
+
+    /**
+     * Returns the distance in miles from the user to the store location.
+     * @param userLocation LatLng of user
+     * @param storeLocation LatLng of store to calculate distance from user
+     * @return distance between the two LatLngs
+     */
+    private double getDistance(LatLng userLocation) {
+      double earthRadius = 3956; // Radius of earth in miles.
+      double distance = 0;
+
+      // Convert latLngs to radians.
+      double userLon = Math.toRadians(userLocation.getLongitude());
+      double userLat = Math.toRadians(userLocation.getLatitude());
+      double storeLon = Math.toRadians(getLongitude());
+      double storeLat = Math.toRadians(getLatitude());
+
+      double deltaLon = userLon - storeLon;
+      double deltaLat = userLat - storeLat;
+
+      // Calculating the distance using the Haversine formula
+      double a = Math.pow(Math.sin(deltaLat / 2), 2) 
+               + Math.cos(userLat) 
+               * Math.cos(storeLat) 
+               * Math.pow(Math.sin(deltaLon / 2), 2);
+
+      double c = 2 * Math.asin(Math.sqrt(a));
+
+      distance = c * earthRadius;
+      return distance;
+
+      // TODO: add error handling
     }
 }
