+++ shopsafe-backend/src/main/java/com/google/sps/servlets/StoresServlet.java
@@ -262,7 +262,6 @@ public class StoresServlet extends HttpServlet {
                     store.getString("formatted_address"),
                     (store.has("opening_hours") && store.getJSONObject("opening_hours").has("open_now")) ? store.getJSONObject("opening_hours").getBoolean("open_now") : null,
                     new LatLng(storeLocation.getDouble("lat"), storeLocation.getDouble("lng")),
-                    location,
                     store.getDouble("rating")));
             }
